* Start work on Heretic cheats * Added BEX mnemonics for Heretic cheat messages * Some cleanup in d_dehtbl.cpp * Change from SGI STL hash to sdbm, which profiled as the second fastest hash given input of the entire English dictionary, and had one of the fewest number of collisions.

+++ source/d_dehtbl.h
@@ -38,8 +38,8 @@ typedef struct dehstr_s
    const char **ppstr;   // doubly indirect pointer to string   
    const char *lookup;   // pointer to lookup string name
    const char *original; // haleyjd 10/08/06: original string
-   int bnext;            // haleyjd: for bex hash chaining (by mnemonic)
-   int dnext;            // haleyjd: for deh hash chaining (by value)
+   size_t bnext;         // haleyjd: for bex hash chaining (by mnemonic)
+   size_t dnext;         // haleyjd: for deh hash chaining (by value)
 } dehstr_t;
 
 extern char **deh_spritenames;
